backend/requirements.txt
# requirements.txt
python-dotenv
requests
google-generativeai
serpapi
fastapi
uvicorn[standard]
google-adk